Only try and poll emails if token is initialised

+++ plugin.rb
@@ -69,29 +69,32 @@ after_initialize do
     def poll_mailbox(process_cb)
         #Initialise token
         self.init_token_from_site_setting()
-      	begin
-          self.refresh_token_if_needed()
 
-          msgraph_api =
-            MsGraphAPI.new(
-              SiteSetting.msgraph_polling_graph_endpoint,
-              SiteSetting.msgraph_polling_mailbox,
-              @token.token
-            )
+	unless @token.nil?
+      	  begin
+            self.refresh_token_if_needed()
 
-          # To avoid managing paging we get all the emails until there are none remaining
-          while (messages = msgraph_api.get_messages_id).length > 0
-            messages.each do |message|
-              mime = msgraph_api.get_message_mime(message)
-              process_cb.call(mime)
-              msgraph_api.delete_message(message)
+            msgraph_api =
+              MsGraphAPI.new(
+                SiteSetting.msgraph_polling_graph_endpoint,
+                SiteSetting.msgraph_polling_mailbox,
+                @token.token
+              )
+
+            # To avoid managing paging we get all the emails until there are none remaining
+            while (messages = msgraph_api.get_messages_id).length > 0
+              messages.each do |message|
+                mime = msgraph_api.get_message_mime(message)
+                process_cb.call(mime)
+                msgraph_api.delete_message(message)
+              end
             end
+          rescue StandardError => e
+            Rails.logger.error(
+              "Error while polling emails with MsGraph plugin: #{e}"
+            )
           end
-        rescue StandardError => e
-          Rails.logger.error(
-            "Error while polling emails with MsGraph plugin: #{e}"
-          )
-        end
+	end
       end
     end
         #If token not initialised then don't run the rest of the polling
